requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
requests==2.31.0
python-dotenv==1.0.1
mysqlclient==2.2.4
django-cors-headers==4.3.1 